+++ crates/modelardb_client/src/error.rs
