+++ crates/modelardb_common/src/error.rs
