+++ crates/modelardb_compression/src/compression.rs
@@ -21,7 +21,7 @@ use arrow::record_batch::RecordBatch;
 use modelardb_types::schemas::COMPRESSED_SCHEMA;
 use modelardb_types::types::{ErrorBound, TimestampArray, ValueArray};
 
-use crate::errors::{ModelarDbCompressionError, Result};
+use crate::error::{ModelarDbCompressionError, Result};
 use crate::models::gorilla::Gorilla;
 use crate::models::{self, timestamps, GORILLA_ID};
 use crate::types::{CompressedSegmentBatchBuilder, CompressedSegmentBuilder, ModelBuilder};
