+++ crates/modelardb_compression/src/error.rs
