+++ crates/modelardb_compression/src/lib.rs
@@ -20,7 +20,7 @@
 #![allow(clippy::too_many_arguments)]
 
 mod compression;
-pub mod errors;
+pub mod error;
 mod merge;
 mod models;
 mod types;
