+++ crates/modelardb_query/src/error.rs
