+++ crates/modelardb_server/src/configuration.rs
@@ -118,7 +118,7 @@ impl ConfigurationManager {
     }
 
     /// Set the new value and update the amount of memory for uncompressed data in the storage
-    /// engine. Returns [`crate::errors::ModelarDbServerError`] if the memory cannot be updated
+    /// engine. Returns [`crate::error::ModelarDbServerError`] if the memory cannot be updated
     /// because a buffer cannot be spilled.
     pub(crate) async fn set_uncompressed_reserved_memory_in_bytes(
         &mut self,
@@ -147,7 +147,7 @@ impl ConfigurationManager {
 
     /// Set the new value and update the amount of memory for compressed data in the storage engine.
     /// If the value was updated, return [`Ok`], otherwise return
-    /// [`crate::errors::ModelarDbServerError`].
+    /// [`crate::error::ModelarDbServerError`].
     pub(crate) async fn set_compressed_reserved_memory_in_bytes(
         &mut self,
         new_compressed_reserved_memory_in_bytes: usize,
@@ -174,7 +174,7 @@ impl ConfigurationManager {
     }
 
     /// Set the new value and update the transfer batch size in the storage engine. If the value was
-    /// updated, return [`Ok`], otherwise return [`crate::errors::ModelarDbServerError`].
+    /// updated, return [`Ok`], otherwise return [`crate::error::ModelarDbServerError`].
     pub(crate) async fn set_transfer_batch_size_in_bytes(
         &mut self,
         new_transfer_batch_size_in_bytes: Option<usize>,
@@ -196,7 +196,7 @@ impl ConfigurationManager {
     }
 
     /// Set the new value and update the transfer time in the storage engine. If the value was
-    /// updated, return [`Ok`], otherwise return [`crate::errors::ModelarDbServerError`].
+    /// updated, return [`Ok`], otherwise return [`crate::error::ModelarDbServerError`].
     pub(crate) async fn set_transfer_time_in_seconds(
         &mut self,
         new_transfer_time_in_seconds: Option<usize>,
