+++ crates/modelardb_server/src/error.rs
