+++ crates/modelardb_types/src/types.rs
@@ -22,7 +22,7 @@
 use std::fmt;
 use std::str::FromStr;
 
-use crate::errors::{ModelarDbError, Result};
+use crate::error::{ModelarDbTypesError, Result};
 
 // Types used for a univariate id.
 pub type UnivariateId = std::primitive::u64;
@@ -77,10 +77,10 @@ pub enum ErrorBound {
 
 impl ErrorBound {
     /// Return an [`ErrorBound::Absolute`] with `value` as its absolute per-value bound. A
-    /// [`ModelarDbError`] is returned if a negative or non-normal value is passed.
+    /// [`ModelarDbTypesError`] is returned if a negative or non-normal value is passed.
     pub fn try_new_absolute(value: f32) -> Result<Self> {
         if !value.is_finite() || value < 0.0 {
-            Err(ModelarDbError::InvalidArgumentError(
+            Err(ModelarDbTypesError::InvalidArgumentError(
                 "An absolute error bound must be a positive finite value.".to_owned(),
             ))
         } else {
@@ -89,10 +89,10 @@ impl ErrorBound {
     }
 
     /// Return an [ErrorBound::Relative`] with `percentage` as its relative per-value bound. A
-    /// [`ModelarDbError`] is returned if a value below 0% or a value above 100% is passed.
+    /// [`ModelarDbTypesError`] is returned if a value below 0% or a value above 100% is passed.
     pub fn try_new_relative(percentage: f32) -> Result<Self> {
         if !(0.0..=100.0).contains(&percentage) {
-            Err(ModelarDbError::InvalidArgumentError(
+            Err(ModelarDbTypesError::InvalidArgumentError(
                 "A relative error bound must be a value from 0.0% to 100.0%.".to_owned(),
             ))
         } else {
@@ -109,13 +109,13 @@ pub enum ServerMode {
 }
 
 impl FromStr for ServerMode {
-    type Err = ModelarDbError;
+    type Err = ModelarDbTypesError;
 
     fn from_str(value: &str) -> Result<Self> {
         match value {
             "cloud" => Ok(ServerMode::Cloud),
             "edge" => Ok(ServerMode::Edge),
-            _ => Err(ModelarDbError::InvalidArgumentError(format!(
+            _ => Err(ModelarDbTypesError::InvalidArgumentError(format!(
                 "'{value}' is not a valid value for ServerMode."
             ))),
         }
